Renamed src and test package from the example
