+++ dob-status-core/src/main/kotlin/com/ytree/dob_fwk/status/core/Main.kt
@@ -1,4 +1,4 @@
-package com.ytree.example
+package com.ytree.dob_fwk.status.core
 
 fun main() {
     val helloWorld = HelloWorld()
